[PL-157] Chore: PlanitCheckbox 컴포넌트화
